remove incorrect comment
There is no call to `readavailable` happening at that point in the code, so it cannot be blocking. The code was in `eof`, and will leave `eof` once the input half of default_stdout is closed and all data is successfully read. We can also make the code more compact by writing directly between the input and output, though this should not be a functional change.